Authentication and JWT setup

+++ backend/src/routes/memories.ts
@@ -1,10 +1,16 @@
 import { FastifyInstance } from "fastify";
 import { prisma } from "../libs/prisma";
-import zod from 'zod';
+import { z } from 'zod';
 export async function memoriesRoutes(app: FastifyInstance) {
-
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+  app.get('/memories', async (request) => {
+    const userId = request.user.sub;
     const memories = await prisma.memory.findMany({
+      where: {
+        userId
+      },
       orderBy: {
         createdAt: 'asc'
       }
@@ -18,37 +24,58 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
-    const paramsSchema = zod.object({ id: zod.string().uuid() })
+  app.get('/memories/:id', async (request, reply) => {
+    const paramsSchema = z.object({ id: z.string().uuid() })
     const { id } = paramsSchema.parse(request.params)
+
     const memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       }
     })
-    return memory
+
+    if (memory.isPublic === true) {
+      return memory;
+    }
+
+    if (memory.userId === request.user.sub) {
+      return memory
+    }
+
+    return reply.status(401).send()
   })
 
   app.post('/memories', async (request) => {
-    const bodySchema = zod.object({ content: zod.string(), coverUrl: zod.string(), isPublic: zod.coerce.boolean().default(false) })
+    const bodySchema = z.object({ content: z.string(), coverUrl: z.string(), isPublic: z.coerce.boolean().default(false) })
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
     const memory = await prisma.memory.create({
       data: {
         content,
         coverUrl,
         isPublic,
-        userId: '1e44235a-3a25-43b8-a7a9-81da997d8ead'
+        userId: request.user.sub
       }
     })
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
-    const paramsSchema = zod.object({ id: zod.string().uuid() })
-    const bodySchema = zod.object({ content: zod.string(), coverUrl: zod.string(), isPublic: zod.coerce.boolean().default(false) })
+  app.put('/memories/:id', async (request, reply) => {
+    const paramsSchema = z.object({ id: z.string().uuid() })
+    const bodySchema = z.object({ content: z.string(), coverUrl: z.string(), isPublic: z.coerce.boolean().default(false) })
     const { id } = paramsSchema.parse(request.params)
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
-    const memory = await prisma.memory.update({
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -62,9 +89,20 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
-    const paramsSchema = zod.object({ id: zod.string().uuid() })
+  app.delete('/memories/:id', async (request, reply) => {
+    const paramsSchema = z.object({ id: z.string().uuid() })
     const { id } = paramsSchema.parse(request.params)
+    
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
     await prisma.memory.delete({
       where: {
         id,
